Add unit tests for dataset manager
